// server/types.ts
export interface IOCItem {
  type: 'IP' | 'DOMAIN' | 'HASH_SHA256' | 'HASH_MD5' | 'FILE_PATH' | 'REGISTRY_KEY' | 'MUTEX' | 'URL' | 'CVE';
  value: string;
  reputation: 'MALICIOUS' | 'SUSPICIOUS' | 'UNKNOWN' | 'BENIGN';
  firstSeen?: string;
  sourceContext?: string;
  geo?: { country: string; city: string; lat: number; lng: number; flag: string };
  threatActor?: string;
  cveData?: {
    cveId: string;
    cvss: number;
    epss: number;
    description: string;
  };
}